[ADD] delete beneficiary RBAC

+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/beneficiary/BeneficiaryDetail.kt
@@ -27,6 +27,7 @@ import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -51,6 +52,7 @@ import com.app.arcabyolimpo.presentation.ui.components.atoms.alerts.DecisionDial
 import com.app.arcabyolimpo.presentation.ui.components.atoms.status.ActiveStatus
 import com.app.arcabyolimpo.presentation.ui.components.atoms.status.InactiveStatus
 import com.app.arcabyolimpo.domain.model.beneficiaries.Beneficiary
+import com.app.arcabyolimpo.presentation.screens.session.SessionViewModel
 import com.app.arcabyolimpo.presentation.ui.components.molecules.NavBar
 import com.app.arcabyolimpo.presentation.ui.components.molecules.TextValue
 import com.app.arcabyolimpo.ui.theme.ArcaByOlimpoTheme
@@ -108,8 +110,11 @@ fun BeneficiaryDetailContent(
     onModifyClick: () -> Unit,
     onDeleteClick: () -> Unit,
     onShowDialog: () -> Unit,
-    onDismissDialog: () -> Unit
+    onDismissDialog: () -> Unit,
+    sessionViewModel: SessionViewModel = hiltViewModel(),
 ) {
+    val role by sessionViewModel.role.collectAsState()
+
     if (showDeleteDialog) {
         DecisionDialog(
             dialogTitle = "¿Seguro que quieres eliminar a ${uiState.beneficiary?.name ?: "..."}?",
@@ -265,10 +270,12 @@ fun BeneficiaryDetailContent(
                         horizontalArrangement = Arrangement.Center,
                         verticalAlignment = Alignment.CenterVertically
                     ) {
-                        DeleteButton(
-                            modifier = Modifier.size(width = 112.dp, height = 40.dp),
-                            onClick = { onShowDialog() }
-                        )
+                        if (role == "COORDINADOR") {
+                            DeleteButton(
+                                modifier = Modifier.size(width = 112.dp, height = 40.dp),
+                                onClick = { onShowDialog() }
+                            )
+                        }
 
                         Spacer(modifier = Modifier.width(16.dp))
